Move GlobalState into preprocess.rs
